Moved LoginHandler to new package. Removed SessionHandler.
